Download button and t411 service download method are bound and work

+++ src/Flan411/Flan411/Tools/T411Service.cs
@@ -71,7 +71,7 @@ namespace Flan411.Tools
             return null;
         }
 
-        public static async Task<string> Download(int torrentId, string fileName)
+        public static async Task<string> Download(long torrentId, string fileName)
         {
             byte[] result;
             byte[] buffer = new byte[4096];
